# swagger.yaml
swagger: "2.0"
info:
  version: 1.0.0
  title: Swagger API Proxy
  license:
    name: MIT
host: petstore.swagger.io
schemes:
  - http
paths:
  /rest/orgUnits:
    get:
      summary: List all orgUnits
      operationId: getOrgUnits
      tags:
        - orgUnits
      parameters:
        - name: where
          in: query
          required: true
          type: string
          maxLength: 250
      responses:
        "200":
          description: 200 success
        default:
          description: unexpected error
    post:
      summary: Create PBX
      operationId: createOrgUnits
      tags:
        - orgUnits
      parameters:
        - name: body
          in: body
          required: true
          schema:
            required:
              - type
              - parentId
              - name
            properties:
              type:
                type: string
                maxLength: 250
              description:
                type: string
                maxLength: 250
              parentId:
                type: integer
                maximum: 9999999999
              name:
                type: string
                maxLength: 250
      responses:
        "200":
          description: 200 success
        default:
          description: unexpected error
  /rest/origUnit/{xId}:
    delete:
      summary: Delete PBX
      operationId: deletePbx
      tags:
        - orgUnits
      parameters:
        - name: xId
          in: path
          required: true
          type: string
          maxLength: 250
      responses:
        "200":
          description: 200 success
        default:
          description: unexpected error
  /rest/orgUnitAttributes:
    post:
      summary: Create max attributes
      operationId: createMaxAttributes
      tags:
        - orgUnitAttributes
      parameters:
        - name: body
          in: body
          required: true
          schema:
            required:
              - name
              - value
              - orgUnitId
            properties:
              name:
                type: string
                maxLength: 250
              value:
                type: integer
                maximum: 9999999999
              orgUnitId:
                type: integer
                maximum: 9999999999
      responses:
        "200":
          description: 200 success
        default:
          description: unexpected error
    get:
      summary: Get properties having in PBX
      operationId: getPropertiesPbx
      tags:
        - orgUnitAttributes
      parameters:
        - name: where
          in: query
          type: integer
          required: true
          maxLength: 10
      responses:
        "200":
          description: 200 success
        default:
          description: unexpected error
  /rest/users:
    post:
      summary: create user
      operationId: createUser
      tags:
        - users
      parameters:
        - name: body
          in: body
          required: true
          schema:
            required:
              - fullName
              - email
              - language
              - orgUnitId
            properties:
              fullName:
                type: string
                maxLength: 250
              email:
                type: string
                maxLength: 250
              language:
                type: string
                maxLength: 250
              orgUnitId:
                type: string
                maxLength: 10
      responses:
        "200":
          description: 200 success
        default:
          description: unexpected error
    get:
      summary: query user manage PBX
      operationId: getAdminId
      tags:
        - users
      parameters:
        - name: where
          in: query
          required: true
          type: string
          maxLength: 250
      responses:
        "200":
          description: 200 success
        default:
          description: unexpected error
  /rest/users/{xId}:
    put:
      summary: block user admin access
      operationId: blockUser
      tags:
        - users
      parameters:
        - name: xId
          in: path
          type: integer
          required: true
        - name: body
          in: body
          required: true
          schema:
            required:
              - id
              - passwordBlocked
            properties:
              id:
                type: integer
              passwordBlocked:
                type: integer
      responses:
        "200":
          description: 200 success
        default:
          description: unexpected error
    delete:
      summary: delete user admin
      operationId: deleteAdmin
      tags:
        - users
      parameters:
        - name: xId
          in: path
          type: integer
          maximum: 9999999999
          required: true
      responses:
        "200":
          description: 200 success
        default:
          description: unexpected error
  /rest/orgUnitAttributes/{xId}:
    put:
      summary: Increase personal extension
      operationId: increasePersonalExtensionId
      tags:
        - orgUnitAttributes
      parameters:
        - name: xId
          in: path
          type: integer
          maxLength: 10
          required: true
        - name: body
          in: body
          required: true
          schema:
            required:
              - id
              - name
              - value
              - orgUnitId
            properties:
              id:
                type: integer
                maximum: 9999999999
              name:
                type: string
                maxLength: 250
              value:
                type: integer
                maximum: 9999999999
              orgUnitId:
                type: integer
                maximum: 9999999999
      responses:
        "200":
          description: 200 success
        default:
          description: unexpected error
    delete:
      summary: delete internal
      operationId: deleteInternal
      tags:
        - orgUnitAttributes
      parameters:
        - name: xId
          in: path
          type: integer
          maximum: 9999999999
          required: true
      responses:
        "200":
          description: 200 success
        default:
          description: unexpected error
  /rest/userRoles:
    post:
      summary: user role
      operationId: userRoles
      tags:
        - userRoles
      parameters:
        - name: body
          in: body
          required: true
          schema:
            required:
              - userId
              - role
              - orgUnitId
            properties:
              userId:
                type: string
              role:
                type: string
              orgUnitId:
                type: integer
      responses:
        "200":
          description: 200 success
        default:
          description: unexpected error
  /rest/addresses:
    get:
      summary: get list Sip
      operationId: getListSip
      tags:
        - addresses
      parameters:
        - name: where
          in: query
          type: string
          maxLength: 250
          required: true
      responses:
        "200":
          description: 200 success
        default:
          description: unexpected error

    post:
      summary: Assign SIP
      operationId: assignSip
      tags:
        - addresses
      parameters:
        - name: body
          in: body
          required: true
          schema:
            required:
              - number
              - orgUnitId
            properties:
              number:
                type: string
                maxLength: 250
              orgUnitId:
                type: integer
                maximum: 9999999999
      responses:
        "200":
          description: 200 success
        default:
          description: unexpected error
  /rest/addresses/{xId}:
    delete:
      summary: Delete Sip
      operationId: deleteSip
      tags:
        - addresses
      parameters:
        - name: xId
          in: path
          required: true
          type: string
          maxLength: 250
      responses:
        "200":
          description: 200 success
        default:
          description: unexpected error
  /rest/validateReg:
    post:
      summary: validate information before register
      operationId: valiedateReg
      tags:
        - register
      parameters:
        - name: body
          in: body
          required: true
          schema:
            required:
              - name
              - email
              - publicNumber
            properties:
              name:
                type: string
              email:
                type: string
                # format: email 
                pattern: '^(([^<>()[\]\.,;:\s@\"]+(\.[^<>()[\]\.,;:\s@\"]+)*)|(\".+\"))@(([^<>()[\]\.,;:\s@\"]+\.)+[^<>()[\]\.,;:\s@\"]{2,})$'
              publicNumber:
                type: string
                pattern: '((84|0)([3|5|7|8|9]))([0-9]{8})\b'
      responses:
        "200":
          description: 200 success
        "400":
          description: Error
        default:
          description: unexpected error
  /rest/validateAddSIP:
    post:
      summary: validate SIP number
      operationId: validateSIP
      tags:
        - validateSip
      parameters:
        - name: body
          in: body
          required: true
          schema:
            required:
              - number
            properties:
              number:
                type: string
                pattern: '((84|0)([3|5|7|8|9]))([0-9]{8})\b'
                maxLength: 250
      responses:
        "200":
          description: 200 success
        default:
          description: unexpected error

definitions:
  Error:
    type: object
    required:
      - code
      - message
    properties:
      code:
        type: integer
        # format: int32
      message:
        type: string